refactor(plugin): extract texture_ring.rs (#1079)
Moves the TextureRingMethodsVTable callbacks +
HOST_TEXTURE_RING_METHODS_VTABLE + accessor + the
texture_ring_methods_vtable_null_tests module out of
host_services/mod.rs into a sibling submodule. mod.rs re-exports
HOST_TEXTURE_RING_METHODS_VTABLE and host_texture_ring_methods_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
